Code/requirements.txt
pycocotools
absl-py==0.9.0
asn1crypto==0.24.0
astunparse==1.6.3
attrs==17.4.0
Automat==0.6.0
blessings==1.7
blinker==1.4
blis==0.4.1
cachetools==4.1.0
catalogue==1.0.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
cloud-init==19.4
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
cycler==0.10.0
cymem==2.0.3
Cython==0.29.19
cytoolz==0.10.1
decorator==4.4.2
distro-info===0.18ubuntu0.18.04.1
docopt==0.6.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.5/en_core_web_sm-2.2.5.tar.gz
et-xmlfile==1.0.1
future==0.18.2
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.2
google-auth==1.15.0
google-auth-oauthlib==0.4.1
google-compute-engine==20190801.0
google-pasta==0.2.0
gpustat==0.6.0
grpcio==1.29.0
h5py==2.10.0
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
importlib-metadata==1.6.0
incremental==16.10.1
jdcal==1.4.1
jellyfish==0.8.2
Jinja2==2.10
joblib==0.15.1
jsonpatch==1.16
jsonpickle==1.4.1
jsonpointer==1.10
jsonschema==2.6.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.2.0
language-selector==0.1
leveldb==0.201
lmdb==0.98
Markdown==3.2.2
MarkupSafe==1.0
matplotlib==3.2.1
mpmath==1.1.0
munch==2.5.0
murmurhash==1.0.2
netifaces==0.10.4
networkx==2.4
nltk==3.5
numpy==1.18.4
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opencv-python==4.2.0.34
openpyxl==3.0.3
opt-einsum==3.2.1
packaging==20.4
PAM==0.4.2
pandas==1.0.3
Pillow==6.1.0
plac==1.1.3
preshed==3.0.2
protobuf==3.12.1
psutil==5.7.0
py-cpuinfo==5.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycrypto==2.6.1
pydotplus==2.0.2
pyemd==0.5.1
pygobject==3.26.1
PyJWT==1.5.3
pymongo==3.10.1
pyOpenSSL==17.5.0
pyparsing==2.4.7
Pyphen==0.9.5
pyserial==3.4
pyspellchecker==0.5.4
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2020.1
pyxdg==0.25
PyYAML==3.12
regex==2020.5.14
requests==2.23.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
rsa==4.0
sacred==0.8.1
scikit-learn==0.23.1
scipy==1.4.1
screen-resolution-extra==0.0.0
seaborn==0.10.1
SecretStorage==2.3.1
service-identity==16.0.0
six==1.15.0
smmap==3.0.4
spacy==2.2.4
srsly==1.0.2
ssh-import-id==5.7
sympy==1.6
systemd-python==234
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
termcolor==1.1.0
textacy==0.10.0
Theano==1.0.4
thinc==7.4.0
threadpoolctl==2.0.0
toolz==0.10.0
torch==1.5.0
torchvision==0.6.0
tqdm==4.46.0
Twisted==17.9.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
wasabi==0.6.0
Werkzeug==1.0.1
wrapt==1.12.1
xkit==0.0.0
xlrd==1.2.0
zipp==3.1.0
zope.interface==4.3.2
